requirements.txt
youtube-transcript-api
streamlit
transformers 
python-dotenv
google-generativeai